+++ infra/tofu/keycloak/roles.tf
@@ -1,0 +1,54 @@
+locals {
+  realm_roles = {
+    "wotbtools-admin" = "WoTBTools admin"
+    "wotbtools-user"  = "WoTBTools user"
+    "boost-manager"   = "Boost manager"
+    "booster"         = "Booster"
+    "HoF-admin"       = "Hall of Fame administrator"
+  }
+}
+
+resource "keycloak_role" "realm" {
+  for_each = local.realm_roles
+
+  realm_id    = keycloak_realm.wotbtools.id
+  name        = each.key
+  description = each.value
+
+  lifecycle {
+    prevent_destroy = true
+  }
+}
+
+resource "keycloak_default_roles" "wotbtools" {
+  realm_id      = keycloak_realm.wotbtools.id
+  default_roles = [keycloak_role.realm["wotbtools-user"].name]
+
+  lifecycle {
+    prevent_destroy = true
+  }
+}
+
+data "keycloak_openid_client" "realm_management" {
+  realm_id  = keycloak_realm.wotbtools.id
+  client_id = "realm-management"
+}
+
+locals {
+  admin_api_realm_management_roles = toset([
+    "manage-users",
+    "query-users",
+    "view-realm",
+  ])
+}
+
+resource "keycloak_openid_client_service_account_role" "admin_api" {
+  for_each = local.admin_api_realm_management_roles
+
+  realm_id                = keycloak_realm.wotbtools.id
+  service_account_user_id = keycloak_openid_client.admin_api.service_account_user_id
+  client_id               = data.keycloak_openid_client.realm_management.id
+  role                    = each.value
+
+  depends_on = [keycloak_openid_client.admin_api]
+}